requirements/requirements-testing.txt
# Pytest for running the tests.
pytest==6.2.2
pytest-django==4.1.0
pytest-cov==2.11.1
codecov==2.1.11
